Created four new classes for the game.
Created Sprite class for future classes to use this as parent class.
Created Player class (that extends Sprite) that defines the characters the player plays with.
Created ClassTypes class that tells the playable characters information about the given in-game-class they are playing as.
Created Ability class that manages the abilities the different in-game-classes will have.